Split types to different documents
