+++ Zilon.Core/CDT.LAST.MonoGameClient/ViewModels/MainScene/BodyPart.cs
@@ -1,0 +1,20 @@
+﻿using Microsoft.Xna.Framework.Graphics;
+
+namespace CDT.LAST.MonoGameClient.ViewModels.MainScene
+{
+    /// <summary>
+    /// Helper structure to keep texture and metadata forbody visualization.
+    /// </summary>
+    public record BodyPart
+    {
+        public BodyPart(BodyPartType type, Texture2D texture)
+        {
+            Type = type;
+            Texture = texture;
+        }
+
+        public Texture2D Texture { get; }
+
+        public BodyPartType Type { get; }
+    }
+}
