+++ Zilon.Core/CDT.LAST.MonoGameClient/ViewModels/MainScene/BodyPartType.cs
@@ -1,0 +1,15 @@
+﻿namespace CDT.LAST.MonoGameClient.ViewModels.MainScene
+{
+    public enum BodyPartType
+    {
+        Undefined = 0,
+        Head,
+        Chest,
+        ArmLeft,
+        ArmRightSimple,
+        ArmRightRifle,
+        ArmRightDual,
+        LegsIdle,
+        LegsCombat
+    }
+}
