+++ Zilon.Core/CDT.LAST.MonoGameClient/ViewModels/MainScene/HandPart.cs
@@ -1,0 +1,20 @@
+﻿using Microsoft.Xna.Framework.Graphics;
+
+namespace CDT.LAST.MonoGameClient.ViewModels.MainScene
+{
+    /// <summary>
+    /// Internal structure to store texture and metadata together.
+    /// </summary>
+    public record HandPart
+    {
+        public HandPart(HandPartType type, Texture2D texture)
+        {
+            Type = type;
+            Texture = texture;
+        }
+
+        public Texture2D Texture { get; }
+
+        public HandPartType Type { get; }
+    }
+}
