+++ Zilon.Core/CDT.LAST.MonoGameClient/ViewModels/MainScene/HandPartType.cs
@@ -1,0 +1,16 @@
+﻿namespace CDT.LAST.MonoGameClient.ViewModels.MainScene
+{
+    /// <summary>
+    /// The part of person hand visualization available on the client.
+    /// </summary>
+    public enum HandPartType
+    {
+        Undefined = 0,
+        Base,
+
+        /// <summary>
+        /// Used for shields in far hand.
+        /// </summary>
+        BaseBack
+    }
+}
